% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00roxygen.R
\name{logit_invcloglog}
\alias{logit_invcloglog}
\title{Logit transformed inverse cloglog}
\arguments{
\item{x}{AD vector}
}
\value{
AD vector
}
\description{
Logit transformed inverse cloglog
}
\details{
Calculates `log(gamma(exp(x)))` with impoved stability for small x.
}
\examples{
F <- RTMB::MakeTape(logit_invcloglog, 1)
}
