% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00roxygen.R
\name{logit_pnorm}
\alias{logit_pnorm}
\title{Logit transformed pnorm}
\arguments{
\item{x}{AD vector}
}
\value{
AD vector
}
\description{
Logit transformed pnorm
}
\details{
Calculates `qlogis(pnorm(x)))` with impoved accuracy for small and large x.
}
\examples{
F <- RTMB::MakeTape(logit_pnorm, 1)
abs(F(-1) - qlogis(pnorm(-1)))
F(-100)
F(100)
}
