% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00roxygen.R
\name{logspace_gamma}
\alias{logspace_gamma}
\title{Log space gamma function}
\arguments{
\item{x}{AD vector}
}
\value{
AD vector
}
\description{
Log space gamma function
}
\details{
Calculates `log(gamma(exp(x)))` with impoved stability for small x.
}
\examples{
F <- RTMB::MakeTape(logspace_gamma, 1)
identical(F(-1), lgamma(exp(-1)))
lgamma(exp(-1000))
F(-1000)
F$jacobian(-1000)
}
